Function for nhanvien

+++ src/main/java/com/herokuapp/realestatebk/service/NhanvienService.java
@@ -6,46 +6,53 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.herokuapp.realestatebk.entity.Nhanvien;
+import com.herokuapp.realestatebk.exception.MessageException;
 import com.herokuapp.realestatebk.form.FormLogin;
+import com.herokuapp.realestatebk.form.FormNhanvien;
+import com.herokuapp.realestatebk.repository.KhachhangRespository;
 import com.herokuapp.realestatebk.repository.NhanvienRepository;
-
 
 @Service
 public class NhanvienService {
-	
+
 	@Autowired
 	private NhanvienRepository nhanvienRepository;
+
+	@Autowired
+	private KhachhangRespository khachhangRespository;
 
 	public List<Nhanvien> getAllNhanvien() {
 		return (List<Nhanvien>) nhanvienRepository.findAll();
 	}
 
-	public void deleteNhanvien(int id) {
-		nhanvienRepository.deleteById(id);;
+	public Nhanvien deleteNhanvien(int id) throws Exception {
+		Nhanvien nhanvienDelete = null;
+		boolean flag = nhanvienRepository.existsById(id);
+		if (flag) {
+			nhanvienDelete = nhanvienRepository.findById(id).get();
+			if (khachhangRespository.countKhachhangByNvID(id) > 0) {
+				throw new Exception(MessageException.messCanNotDeleteNhanvienHasKhachHang);
+			} else {
+				nhanvienRepository.deleteById(id);
+				return nhanvienDelete;
+			}
+		} else {
+			throw new Exception(MessageException.messNhanvienNotExists);
+		}
 	}
 
-	public Nhanvien addNhanvien(Nhanvien nhanvien) {
-		Nhanvien nhanvienAdd = nhanvienRepository.save(nhanvien);
+	public Nhanvien addNhanvien(FormNhanvien fNhanvien) {
+		Nhanvien nhanvienAdd = nhanvienRepository.save(fNhanvien.coverToNhanvien());
 		return nhanvienAdd;
 	}
 
-	public Nhanvien editNhanvien(Nhanvien nhanvien) {
+	public Nhanvien editNhanvien(FormNhanvien fNhanvien) {
 		Nhanvien nhanvienEdit = null;
-		boolean flag = nhanvienRepository.existsById(nhanvien.getNvid());
-		if(flag) {
-			nhanvienEdit = nhanvienRepository.save(nhanvien);
+		boolean flag = nhanvienRepository.existsById(fNhanvien.getNvid());
+		if (flag) {
+			nhanvienEdit = nhanvienRepository.save(fNhanvien.coverToNhanvien());
 		}
 		return nhanvienEdit;
-	}
-	
-	public Nhanvien checkExistNhanvien(String email, String pass) {
-		Nhanvien checkedNhanvien = null;
-		List<Nhanvien> listNhanvien = (List<Nhanvien>) nhanvienRepository.findAll();
-		for (Nhanvien nv : listNhanvien) {
-			if(email.equals(nv.getEmail()) && pass.equals(nv.getMatkhau()))
-				checkedNhanvien = nv;
-		}
-		return checkedNhanvien;
 	}
 
 	public Nhanvien login(FormLogin formLogin) {
